MSVC: Use "/external:I" flag to set the system include paths
Previously we were using the "/I" compiler option to specify
the system include paths, because the MSVC compiler had not
the special option for that.

But since v15.6, the MSVC 2017 compiler has the special
"/external:I" option to specify the system include paths:

 * https://devblogs.microsoft.com/cppblog/broken-warnings-theory/

So, it makes sense to use this new option for the MSVC versions
which are supports it (since _MSC_VER 1913, aka MSVC 2017
update 6, v15. 6.0):

 * https://dev.to/yumetodo/list-of-mscver-and-mscfullver-8nd

But, due to MSVC 2017 update 6 has some bugs:

 * https://developercommunity.visualstudio.com/content/problem/181006/externali-include-paths-not-working.html

which are fixed since MSVC 2017 update 9 (aka _MSC_VER 1916),
then we just keep _MSC_VER 1916 as a base version.

In other cases we fall back to the "/I" option.

Fixes: QBS-1573
Change-Id: Ie180ba61a1988aaec7955f545b989edf406c7730
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>