Suppress leaks from QThreadPrivate()
It looks like there's a leak in Qt code - the QThreadData object is
leaking in some obscure cases (or the bug in Qbs is too tricky to be
spotted in reasonable time), so suppress those leaks for now.
Note that QArrayData::allocate leaks from the posted event list object
which is stored in QThreadData too, so this is the same leak.

Change-Id: I64a2cce1746ce5565dc79b5a142c7c146e1a7391
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>