baremetal: Fix linker name detection for ARM using IAR toolchain
This commit ab0abfc introduced a regression in the detection of
a linker name for the ARM architecture.

A problem is that now the ARM architecture has a detailed name,
e.g. "armv7", "armv6", against the previous "arm" name. In this
case the following code:

 "ilink" + architecture

returns a wrong linker name, e.g. "ilinkarmv7" instead of "ilinkarm".

So, we need to check on ARM architecture and then return a valid
linker name directly.

Change-Id: I4f44cdc9875f00676bc36cfa4fc36b36b92a8190
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>