baremetal: Derive all applications and libraries from common products
The problem is that for different architectures and different toolchains
we need in a minimal set of a platform-specific flags and etc.

Previously (e.g. for MCS51 architecture and IAR/KEIL/SDCC toolchains) we
are not faced with this problem, because there by default we don't need
to specify a platform-specific flags for a minimal test.

But, for other platforms and compilers we need in this basic flags
(e.g. for ARM architecture and KEIL toolchain, we need to set the CPU
target flag, e.g. '--cpu cortex-m0').

So, it makes sense to derive all tested baremetal applications and
libraries from the basic BareMetal{Application|StaticLibrary} products.
In this case we can configure the desired flags once, using the QBS
Properties item.

Tested with the following architectures and toolchains:

 * MCS51 (KEIL, IAR, SDCC)
 * STM8 (SDCC)
 * ARM (KEIL)

Change-Id: I484ed3abf78050fe2dd27e96e9442a3a03cecfa5
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>