baremetal: Allow libraries linking with SDCC on Windows host
This commit implements a workaround for the SDCC bug on a Windows host:

 * https://sourceforge.net/p/sdcc/bugs/2970/

A bug is in that the SDCC generates an object files with the DOS line
endings '\r\n' instead of the UNIX line endings '\n' on a Windows host,
that leads to the linking errors.

We need to patch the generated object files to use only the UNIX line
endings.

Change-Id: I055169d0e707b49434af3a7a8c488ae597d1c889
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>