baremetal: Simplify cpp.prefixHeaders handling using KEIL ARM toolchains
It seems it makes no sense to add the 'pre-include' flags before the
'defines' and 'includes', because all pre-includes will be pre-processed
after the 'defines' and 'includes'.

At least this note belongs to the 'armclang' compiler:

 * http://www.keil.com/support/man/docs/armclang_ref/armclang_ref_sam1452772182261.htm

So, we can simplify the cpp.prefixHeaders handling to minimize the
JS code duplication.

Change-Id: I7c6dd7b5545294ff6aba866c2be9ebbcff190f6f
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>