Do not rely on PathProbe returning relative paths
This testcase uses relative paths both as input properties to PathProbe
and as expected results (rhs). The obtained output properties (lhs) are
relative paths, too and as such, a simple string comparison is performed
to verify the results. While this currently works, it is not correct to
do so.

PathProbe features input and output properties of pathList type and
thus, when reading these properties, returned values should always be
absolute file paths. It is a known problem that path properties in
probes are not absolutified, which will be fixed when porting to
QJSEngine. Thus, we need to absolutify the expected values when
evaluating the results.

In order to remain compatible with the current implementation, we need
to do the absolutification on both sides.

Change-Id: I13c7655f509179f3778c736e08fbd1bd264a63c4
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>