baremetal: Introduce new 'none' choosing for qbs.targetPlatform
This 'none' value should be used for a bare-metal toolchains
in the qbs.targetPlatform and the qbs.targetOS properties.

Now this value is set in the profiles during automatic detection
of such toolchains as IAR, KEIL and SDCC.

Take into account, that it is impossible to do the same and for
the GCC toolchain, because there are not exists a proper way to
know that a current GCC is for a 'bare-metal'. So, the users
should change the qbs.targetPlatform manually.

Change-Id: I178fd526bbb5ae44f18e35ff12970ed0718c476d
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>