baremetal: Fix static library generation for GCC compiler
A problem was in that the GenericGCC.qbs file missing the
initialization for the staticLibrary{Prefix|Suffix} properties.

This properties are initialized only in the derived
UnixGCC.qbs or MingwBaseModule.qbs modules, which are does
work only for the 'unix' or 'windows' platforms.

So, we need to initialize required properties inside
of a base GenericGCC.qbs module.

Change-Id: Ie25f1a19acbc5d86e65eb5ac8f8205ae8079ebd9
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>