baremetal: Pass 'compiler-listing-files' test for SDCC toolchain
A problem was in that the SDCC compiler does not have an option
to disable generation for listing files. Besides, it use listing
files for a linking.

So, we need to use a workaround to remove the generated listing files
after linking in case if the cpp.generateCompilerListingFiles
property is false.

Change-Id: I5913e62d307d296bd8891654d20787cf3a54833c
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>