baremetal: Fix detection of ARM architecture details for KEIL armclang
The '__ARM_ARCH_PROFILE' macro contains the desired ARM profile
letter, which appends to the resulting ARM architecture name.

But a problem is that this letter wrapped in quotes, e.g. 'M'.
In this case the constructed architecture also has this quotes,
e.g. armv6'm', that's wrong.

We need to remove all quotes from the found letter and then
append this letter to the resulting string.

Change-Id: Idea61082ebb584151ddcb1acc9eba62a516ea170
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>