Avoid wrong-typed temporary assignment to probe property
The code assigns a temporary string to the cflags and libs properties
which are both declared as stringList. Later, it reads these properties,
expects string values and converts them into arrays.

While this seems to work with the current QtScript engine, it won't with
the QML-based evaluation engine. That is because any read access to a
item property will return exactly the declared type or undefined. In
this particular case, we would assign the temporary string to a
stringList property which the engine would automatically convert into a
stringList (array) when reading and thus, calling split() would result
in an error.

Change-Id: I53e0c51ee042f787e0d444cdc6cfe4b8820e46ac
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>