GitHub actions: Fix caching on Windows (v2)
There is no way to manually clear the cache and we have an entry
"Windows-clcache-" stuck in cache which is exact match for the restore-
keys and thus it has a higher priority than the key with timestamp.
Fix that by changing key used on Windows.

Also, simplify "prepare timestamp" steps

This amends 3dcfba17b.

Change-Id: I10bdd6dd3e4607870de49facdfcb43e842579552
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>