Re-create QML parser from grammar file
This has not been done for years. In the meantime the generated code has
been modified without backporting the changes to the grammar file. In
preparation of potential future updates of the QML grammar, we need to
make sure that we can reproduce the parser files and minimize manual
post-processing.

Small manual changes to qmljsgrammar_p/.cpp (license and namespace) are
still necessary because unlike for qmljsparser_p.h, the template is
hard-coded in qlalr.

Change-Id: Ie920d4c0ca319a7d07bb83b0f3944531a4d7b7fa
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>