Android: Remove targetName suffix from QtApplication
From Qt 5.14 target suffix is required for the multi arch packaging. This only affects
android applications using Qt. The suffix can't be applied in QtApplication because
a project using CppApplication and depending on Qt.core won't have it.
Beside using targetName is error prone as the user can change it.
So instead of using targetName for the suffix this patch adds the suffix when
deploying the "android.nativelibrary" for packaging and only for Qt applications.

Fixes: QBS-1608
Change-Id: I0704d5433c8a1d513dfdef71cecbc99ea5458181
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>