Disable moc for the aggregate product when multiplexing
Running moc in the aggregate can't happen anymore because it generates
an error when accessing cpp.defines.
Before 527d5a8ec54e44291d4968754bdf574ebe57822b cpp.defines was not
accessed for hpp files. Anyway it doesn't make sense to call moc
in the aggregate as the generated files won't be compiled.

Change-Id: I25ada3c8c54b199efb0877b59025fff88dc84dd8
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>