JSON API: use fullDisplayName in dependencies array
Change the dependencies array to use the product full display name
instead of just the name. That allows clients to search for the
multiplexed products using the dependencies json array of the aggregate
product. Clients are then able to properly update the search path of
the dynamic libraries generated in the multiplexed products.

Change-Id: I3dd8b66e2cb637e501cbfe29c670c57c2cea5f7b
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>