Android: Fix debugging support
Add archSuffix to generate the library with the architecture name for
Qt > 5.14 (before the lib was renamed which prevented debugging).

Remove rpaths as they are not used on android and generate warnings.

Update target and defines to comply with android's Build System
Maintainers Guide.

Remove stripping step in Qt.android_support as androiddeployqt
actually does the stripping.

Change-Id: I46a273c1cded753223c7b759670477815b625267
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>