Do not modify the product object
Since the script creates and modifies a local 'imports' variable, the
local variable should also be used for iteration. Otherwise we rely
on 'imports' being a reference only. While this is the way how copying
objects works in JS, it is still not clean, especially since we view
the product object as read-only.

Change-Id: I1ffe61e03b793d5ed36d069e983014172db85463
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>