baremetal: Improve detection of IAR compiler includes
The IAR compiler has an undocumented command line option `--IDE3`,
which allows you to print a list of compiler include paths as:

`$$TOOL_BEGIN $$VERSION "3" $$INC_BEGIN $$FILEPATH "<path\\to\\directory>" $$TOOL_END`

Besides, the same approach is used in the IAR extension for
the VSCode IDE.

So we can use this approach to implement it the Qbs module as well.

In addition, this commit contains an autotest to check the
`cpp.compilerIncludePaths` property.

Change-Id: I434dd630913e5afd6cba5b4e31e1021ee0c5fe31
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>