baremetal: Return only existing SDCC compiler include paths
For some architectures, the compiler may return nonexistent
header paths. For example, for the `stm8` architecture, it
returns the path:

 c:\Program Files (x86)\SDCC\bin\..\include\stm8

which is not present on the file system.

This can be confusing intellisense for IDEs, and will also
make it harder to write tests for `cpp.compilerIncludePaths`
property.

So, it makes sense to return only existing paths.

Change-Id: Ibe7078485db900eba2e17710e5de0abeda28f645
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>