Get rid of product.moduleProperty() during project resolving
Although not documented, it was possible to access module properties via
product.moduleProperty("module", "property"). That is still possible in
the QJS port of Qbs, but would complicate the implementation of Evaluator
unnecessarily. After all it was only needed in a single place. It can
still be used in rules.

Change-Id: I40f374a57a6fefc85eb45250b472fe732b66f0b7
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>