baremetal: Patch *.rel files generated by SDCC compiler on windows host
... before linking.

SDCC still has an bug on Windows where it generates the object files
in the DOS format instead of Unix format, that leads to the linking
errors:

 * https://sourceforge.net/p/sdcc/bugs/2970/

A workaround is to remove all CR characters from the object files,
what was done in the previous fix in Qbs. But that solution does not
work since the TextFile service forcibly adds the CP characters to
the ends of lines on the Windows host. We need to use the BinaryFile
service instead which operates with the RAW data.

Change-Id: I09f44a707081de6af7300348e743b6076366dd2c
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>