baremetal: Fix detection of GCC ARM toolchain installers on Windows
... provided by ARM group.

Previously, we detected the installed GCC toolchain through the
`Uninstall` group in the Windows registry. Where we have enumerated
all sub-groups in this group, and searched for a matching for to the
`GNU Tools for ARM Embedded Processors` token. But the problem is
that in newer versions of the installer, the name of this group has
been changed. Moreover, that method is slow because it enumerates
all the entries in the group, where these entries can be a lot.

Alternatively, we can search for installed toolchains using the
`HKLM/SOFTWARE/ARM` path that is constant, contain only the installed
toolchains, and present on all versions of installers.

Change-Id: If3dc280d03a9c6e25113263309822599f0010d9b
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>