Share cpp::objectSuffix property
It makes sense to add the cpp.objectSuffix property to the base
CppModule due the following reasons:

1. It is possible that the user wants to change the extension for
the generated object files, which makes working with Qbs more flexible.

2. It will be easier to write an autotests that check the generation
of the object files for a bare metal platforms, where object files can
have various extensions such as ".rel", ".r90", ".r43", and so forth.

Change-Id: Ib73836d5e614a6924d312d061f02437c3b318785
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>