Share cpp::linkerMapSuffix property
It makes sense to add the cpp.linkerMapSuffix property to the base
CppModule due the following reasons:

1. It is possible that the user wants to change the extension for
the generated map files, which makes working with Qbs more flexible.

2. It will be easier to write an autotests that check the generation
of the map files for a bare metal platforms, where object files can
have various extensions such as ".map", ".m51", ".m66", and so forth.

Change-Id: I1e5e3b39bf84fddfe8c06f2db9d3936c6a3ff027
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>