baremetal: Improve linkerMapFile() test
Right now we can use the cpp.linkerMapSuffix property
to find out the currently used linker map suffix.

Also we can set a custom cpp.linkerMapSuffix value to
make sure that this applies to the module.

In addition, the generation of a custom linker map file
for SDCC toolchain has been fixed.

Change-Id: I8798cd6bea0ab6b5ea9728400827b8c98b11ba7b
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>