Fix detecting MSVC via Probe when multiple versions are present
Previously, Qbs iterated over different versions in the ascending order.
During the setup-toolchains this means that the newest one will be
written in the settings as only the last one is actually written.

When running ClBinaryProbe, the first one (i.e. the oldest was picked
up) which did not work well with vcvarsall (without -vcvars_ver
parameter, it uses the newest one). So, pick up the newest compiler both
when running setup-toolchains and when detecting via Probe.

Task-number: QBS-1498
Change-Id: Ib1b433ca7e17747dee986ba383a3c01ee91851fb
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>