Fix reading Xcode 12.5 specs
Since Xcode 12.5, common specs for darwin platforms were moved to
"Darwin Package/Product Types.xcspec" in the different folder. However,
the old configs are still present, so we need to merge info from both
variants.
Also, allow to use undefined variables in .xcspec files - since 12.5,
this is a valid use-case.
Also, implement the "default=" variable formatter used as a fallback
when variable is not set.

Fixes: QBS-1644
Change-Id: I9a55c5d511cbb01f30219478ee9f3b85763a08d5
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>