baremetal: Long live COSMIC toolchain for STM32 ARM architecture
This patch adds support for Cosmic Software compiler for the
bare metal platforms:

 * https://www.cosmicsoftware.com/download.php

Note that this compiler only works on Windows host platforms.

Right now this patch implements support for ARM architecture for
STM32 microcontrollers. But in the future it is simple to extend
it and for other architectures (e.g. for STM8, HC08 and so on).

Also the COSMIC toolchain has been added to the Github CI.

Change-Id: Iabdfd61cc427155117730db46e067004ceb07102
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>