Fix using extraSearchPaths stack
We push productContext->searchPaths via SearchPathsManager in the
setupProductDependencies() and then add some paths via providers. This
results to the fact that SearchPathsManager pops providers path but not
the one added in setupProductDependencies().

Fix that by restoring the state completely where it was before
resolveDependencies() call. This is fine, since providers paths are also
added to productContext->searchPaths and second Dependency resolution
pass will use it.

This allows to get rid of the newlyAddedModuleProviderSearchPaths which
seems to be needed only to count elemets in the stack correctly.

This is implemented be remembering the extraSearchPathsStack.size() in
SearchPathsManager object.

Change-Id: I8f2352e652d27f2a0b770c2a026f115e5a4ded9a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>