Move module providers code to the separate class
ModuleLoader is too big and more logic will be added to module
providers, so it make sense to extract some code to a separate class.

Unfortunately, it is hard to break the dependency between types
completely - it is tempting to pass ModuleLoader::ProductContext into
new class functions. Alternative would be to pass all necessary data via
function parameters which will make the code less readable.

Change-Id: Ida61192348ef7db89b21f0d58f05e61969e2d01c
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>