codesign: Fix checking if product is a framework
It is not correct to check for the frameworkVersion to determine if
product is a Framework - e.g. Xcode always sets this variable. Instead,
product.type or product.bundle.packageType should be checked instead.

This fixes the problem when codesign incorrectly determines bundle as a
framework if product.version is set.

Fixes: QBS-1649
Change-Id: I067e852df82a43897000d45e27908dd5e22eac70
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>