Do not modify the global state after running provider
Instead, cache the result of the provider in a QHash using the provider/
module name and config as the key. This allows to have a clean state for
each Depends item, so the order in which they appear does not influence
which modules are generated. This is required for the later patch that
implements "named" providers.

Change-Id: Ia395cc94430763ed33d7ff5f2ee39e36d64f195e
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>