Long live Digital Mars toolchain
This patch adds an experimental support for the Digital Mars
C/C++ compiler for Windows:

 * https://www.digitalmars.com/

This compiler is only distributed for Windows host platform,
but allows to create the binaries for Win32, Win16, DOS32 and
DOS targets.

This patch implements only the Win32 targets and has the
following limitations:

 * Support for precompiled headers is not implemented yet
(but which is supported by the compiler).

 * There is no support for the CI autotests due to the fact
that the compiler has the following restrictions:
  - it needs a separate masm386 assembler.
  - it needs a special format of DLL library sources (required
    the DLL main function).
  - it needs the `kernel32.lib` to be always explicitly linked.
  - and, maybe other unexpected stuff.

The following features are currently supported:

 * Compiling a source files in the C or the C++ mode.
 * Compiling the source resource (RC) files.
 * Creation of an applications (both console and GUI).
 * Creation of an dynamic (DLL) libraries with its import files.
 * Creation of a static libraries.

Task-number: QBS-1598
Change-Id: I5a6ac05a7e7513a1c6d2403db037c309b0f12b90
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>