CMake build: Fix build with Qt Creator and CMake < 3.18
Cannot create alias targets for targets that are imported.
For CMake >= 3.11 we could try to force the IMPORTED_GLOBAL
property like done with Core5Compat, but in practice this
could lead to other issues and would not be supported with
CMake 3.10.

Make qbsscriptengine an interface library that links to
Qt5::Script instead.

Amends ea0e25b4ca6adeebe33187e5c12dd254aaa55fec

Change-Id: Id50475a7f9bdad4566bde760806539b990c72add
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>