Use common Cpp.resourceCompilerOutput{Tags|Artifacts} functions
... to unify a codebase.

Resource output artifacts are now tagged with the 'res' instead of
the 'obj' as before (also the output file extensions are also changed
from '*.obj' to '*.res'). This only affected to the additional
processing for the new 'res' tags in the linker rules for the MSVC
and the MinGW.

The reason for adding the new 'res' tag is that for some compilers
(e.g. the future Open Watcom), the resource files must be linked using
the special flags (and not passed as like regular object files to the
linker).

Change-Id: Ibfc52a384816d3769a7c1f6564db079c118011a0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>