baremetal: Get rid of relative paths where it is possible for COSMIC
The COSMIC linker has some restrictions on the length of the object
paths, and it reports the error, like this:

 ERROR: #error clnk cosmic.lkf:1 object pathname too long: (...)

Also the problem is that the COSMIC toolchain does not support the
response files. So, we used the workaround, where for each command
we set the working directory to the build directory, and instead of
the full paths to the files, we used the relative paths for all rules.

But this limitation makes it impossible to parse the compilation
warnings or errors at integrating with the various IDEs, because
the paths to the files are not absolute (e.g. integration with the
'problem' panes or something else).

Right now it is enough to leave the relative paths only when
specifying the paths to the libraries in the linker rule, while the
rest of the other paths are made absolute.

Change-Id: I159bf4e8163e5f9d0aaa1e2555df3dbd81f390c5
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>