pkgconfig: Implement early dependency merging
Due to performance reasons, we have to do the dependency resolution
on the qbspkg-config side, like original pkg-config does.
Otherwise, it is not possible to use GRPC module as it depends on a
bunch of ABSEIL libraries which has a lot of cross-dependencies and
QBS cannot handle them (resolving takes > 10 minutes for a simple
dependency on GRPC).

Change-Id: Ice25f6cea09fe731a511d91891936ba746e5b697
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>