Android: correct reading of prl files and remove pthread lib
prl files for Qt versions 6.0 and 6.1 don't have an architecture
suffix.
prl files from Qt version 6.0 require pthread in the QMAKE_PRL_LIBS
variable but the pthread functionality is included in the libc on
android.
The issue did not trigger an error before Qt 6.2 because the prl file
was not read.

Change-Id: Ia59d1102d6136aa87bfc17df7c6981048b0d0fca
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>