Long live Open Watcom toolchain
This patch adds basic support for the Open Watcom toolchain.
This patch uses the `owcc` compiler (supplied with the toolchain),
which is a wrapper that supports the POSIX standard.

Reason is that the native OW compiler and linker has a limitations
in the command line arguments (e.g. they have wrong quotes handling
and so on).

This patch supports both the latest official version v1.9 and also
its fork v2.0.

Also added the CI autotests for the version v2.0 for the Windows
host. These autotests only perform a limited number of tests (only
the bare-metal tests) due to the following toolchain limitations:

 * The toolchain does not have STL support (there seems to be some
kind of the partial support in the form of separate legacy STL ports).

 * The toolchain support something compatible with the C++98 standard.

 * The toolchain does not support the shared libraries on Linux hosts.

These limitations make it impossible or unjustified to reuse most
of the available tests (it requires a lot of work).

There was also an attempt to set up CI for tests on Linux host, but
for some reason the toolchain installer crashes on CI (although it
works fine on a local PC with Ubuntu).

Change-Id: Iecf76f51f0b09d31a89683f786b9cd7a825f235e
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>