Fix detection of MSVC environment for cross-compilers
The MSVC profile has no mention of the host architecture, since it
contains only the target architecture.

Therefore, in the probe for the MSVC cross-compilers such as HostX64\x86,
HostX86\x64, only the target architecture is passed to determine their
build environment, which is incorrect.

We need to pass the combined cross-architecture as x86_x64, x64_x86 by
analogy as it is done in the qbs-setup-toolchains.

Fixes: QBS-1681
Change-Id: I549d8673cd6d4b0e46bf714355206b9efb2f5b1f
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>