Import span from QtCreator
This is an implementation of the C++20 std::span class [0].

Span is similar to std::string_view and is used as a view type
for vectors; however, unlike string_view, it allows mutable views
(i.e. allows to modify elements of a vector for a non-const T).

Since C++20 is not coming anytime soon on all supported platforms,
it makes sense to import a generic implementation.

We use the lite implementation by Martin Moene [1].

[0] https://en.cppreference.com/w/cpp/container/span
[1] https://github.com/martinmoene/span-lite

Change-Id: Ic53fab81ceb5e9c841111f1efa248152f14e5a37
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>