Add support for other platforms supported by Digital Mars
This patch adds support for the following target platforms:

 * DOS - both 32/16 bit.
 * Windows - 16 bit.

Also, more specifically, the target is controlled by the
`cpp.extenderName` property, which is the DOS extender to use.

The Digital Mars support the following DOS extenders:

 * Rational 16 bit DOS Extender
 * ZPM 16 bit DOS Extender
 * DOSX 32 bit DOS Extender
 * Pharlap 32 bit DOS Extender

which have the following names `dosr`, `dosz`, `dosx`, `dosp`.

The user can set any of the desired values in the `cpp.extenderName`
property (also this property can be empty that means no any
extender is used).

Right now the `qbs-setup-toolchain` utility creates all possible
supported profiles with the required pre-configured properties:

 Profile 'dmc-8_57_0-windows-x86' created for 'C:/dm/bin/dmc.exe'.
 Profile 'dmc-8_57_0-windows-x86_16' created for 'C:/dm/bin/dmc.exe'.
 Profile 'dmc-8_57_0-dosx-x86' created for 'C:/dm/bin/dmc.exe'.
 Profile 'dmc-8_57_0-dosp-x86' created for 'C:/dm/bin/dmc.exe'.
 Profile 'dmc-8_57_0-dosr-x86_16' created for 'C:/dm/bin/dmc.exe'.
 Profile 'dmc-8_57_0-dosz-x86_16' created for 'C:/dm/bin/dmc.exe'.
 Profile 'dmc-8_57_0-dos-x86_16' created for 'C:/dm/bin/dmc.exe'.

Compiling with DOS extenders requires additional external utilities
and libraries (depending on the extender type). Some of them are no
longer available at the moment, so the tests in CI are only done for
the `dmc-windows-x86`, `dmc-windows-x86_16`, and `dmc-dos-x86_16`
profiles.

Change-Id: I221d2995900c63605c8552b825bf0c4d55171f5c
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>