baremetal: Add launch of test application target if possible
The idea is to check that the compiled application works correctly
by running it and reading its output. This will only work if the host
platform and target platform are the same, otherwise the launch is
simply skipped.

Since we use baremetal tests also to test other platforms (for example,
using the OpenWatcom or DigitalMars toolchains), the base application
`BareMetalApplication.qbs` is created as a console application.

The reason is that for baremetal it is always a console, but for
non-baremetal it can also be a GUI which requires a different main
function.

Change-Id: I520ba411582223950521b4b06c2e01bce49fe089
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>