baremetal: Add test to check toolchain probes
Right now we don't have tests that could check the correctness of the
properties returned by toolchain probes.

For example, such as `architecture`, `targetPlatform`, `endianness`
and so forth. The number of returned parameters depends on the type
of toolchain.

This patch adds a test for the probe for the `watcom` toolchain and
can be used as a template for adding the tests for other probes in the
future.

Change-Id: I5ebda0deb7538f746406f5b0925bdd60fdb4c3fb
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>