baremetal: Add simple test for linking with shared libraries
As we use the baremetal tests for non-baremetal compilers
(for example, for `watcom` and `dmc`), then we need a test for
linking the application and the shared library.

We cannot use the existing tests for `watcom` and `dmc` toolchains,
because these toolchains have specific behavior and also do not
fully support the STL. Therefore, the simplest dependency linking
test is used here.

Also fixed the shared library creation for these toolchains.

Change-Id: I0e5d5ede39fa0c9b4bf7db54adc3f161e0aea91c
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>