Switch JavaScript back-end
Newer clang versions seem to expose serious bugs in QtScript, whose
complexity makes it difficult to track them down.
We therefore switch to the more light-weight QuickJS, which offers all
the features we need (most notably property access interception), as
well as good performance.

To save some porting effort, we removed the long-deprecated loadFile()
and loadExtension() functions.
During the porting procedure, we noticed and fixed thread safety issues
in artifact access from JS commands.
We consider this change important enough to bump the major version, so
the next release will be 2.0.

Detailed benchmarking data is below. In summary, we see a modest speed-
up at the cost of a similarly modest increase in memory consumption
(with the exception of project resolving on macOS, which has become a
bit slower). Importantly, the increase does not rise with project size,
as the comparison of qbs vs Qt Creator shows.

Output of qbs_benchmarker on Linux with qbs as test project:
========== Performance data for Resolving ==========
    Old instruction count: 12870602895
    New instruction count: 11923459780
    Relative change: -8 %
    Old peak memory usage: 61775848 Bytes
    New peak memory usage: 67583424 Bytes
    Relative change: +9 %
========== Performance data for Rule Execution ==========
    Old instruction count: 4074062223
    New instruction count: 3887473574
    Relative change: -5 %
    Old peak memory usage: 35123704 Bytes
    New peak memory usage: 38398392 Bytes
    Relative change: +9 %
========== Performance data for Null Build ==========
    Old instruction count: 1104417596
    New instruction count: 1011033948
    Relative change: -9 %
    Old peak memory usage: 24461824 Bytes
    New peak memory usage: 25325920 Bytes
    Relative change: +3 %

Output of qbs_benchmarker on Linux with Qt Creator as test project:
========== Performance data for Resolving ==========
    Old instruction count: 67166450352
    New instruction count: 60772791018
    Relative change: -10 %
    Old peak memory usage: 327011616 Bytes
    New peak memory usage: 343724176 Bytes
    Relative change: +5 %
========== Performance data for Rule Execution ==========
    Old instruction count: 71684351183
    New instruction count: 67051936965
    Relative change: -7 %
    Old peak memory usage: 374913688 Bytes
    New peak memory usage: 387790992 Bytes
    Relative change: +3 %
========== Performance data for Null Build ==========
    Old instruction count: 8383156078
    New instruction count: 7930705668
    Relative change: -6 %
    Old peak memory usage: 180468360 Bytes
    New peak memory usage: 182490384 Bytes
    Relative change: +1 %

Real-world data building Qt Creator (using qbs --log-time, several runs,
removing outliers):
  macOS:
    Resolving: 43s -> 47s
    Rule execution: 17s -> 14s
  Windows:
    Resolving: 18s -> 16s
    Rule execution: 22s -> 17s

Fixes: QBS-913
Fixes: QBS-1103
Fixes: QBS-1126
Fixes: QBS-1227
Fixes: QBS-1684

Change-Id: Ie5088155026e85bbd1e303f1c67addb15810a3cb
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>