Fix Android build-tools detection, if it was installed via cmdline-tools
The modern way to install android packages is to use cmdline-tools. The
old way was to use SDK tools, but now it is deprecated (see
https://developer.android.com/studio/releases/sdk-tools). There is a code
in AndroidSdkProbe which looks for android executable in SDK tools folder
to prove, that the folder does contain Android SDK. But if you use
cmdline-tools, there is no folder "tools" anymore. The probe fails to find
build tools. The fix is to look for build-tools directory.

Fixes: QBS-1700
Change-Id: Iad92664ba07d034ba365e8b7b178cd47f5d3f187
Reviewed-by: Raphaël Cotty <raphael.cotty@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>