cmake: fix installing documentation
The HTML documentation was not installed if QCH doc was not
requested, fix that.

Introduce a new 'qbs_docs' target which should be used instead
of the 'BuildQbsDocumentation' target. This is done for
compatibility with QMake.

Also, introduce a single 'qbs_docs' component for installing docs
in one go - this is required for MacPorts since there is no easy way
to install multiple components at once with CMake.

Which docs will be installed is controlled at the configure step:

$ cmake -DQBS_INSTALL_HTML_DOCS=1 -DQBS_INSTALL_QCH_DOCS=1 -DQBS_INSTALL_MAN_PAGE=1
$ cmake --build . --target qbs_docs # not required for man-page only install
$ cmake --install . --component qbs_docs

Change-Id: I174662c54255a8fb10529b1961c298eb1f1b1470
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>