GitHub actions: select default xcode before running tests
Some tests, such as TestBlackbox::minimumSystemVersion or
TestBlackbox::emptyProfile use the default Xcode version
instead of the one specified by the profile.

This leads to fails of those tests with very old/new
xcode versions.

Fix that by making sure that the default Xcode is the
same as the one specified via profile.

Change-Id: I6d741ece7924d836c1fc66b496fe135bb47de72e
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>